rdkafka_performance: fix size_t warning

+++ examples/rdkafka_performance.c
@@ -1166,8 +1166,8 @@ int main (int argc, char **argv) {
 			RD_KAFKA_PARTITION_UA;
 
                 if (latency_mode) {
-                        size_t minlen = (int)(strlen("LATENCY:") +
-                                        strlen("18446744073709551615 ")+1);
+                        int minlen = (int)(strlen("LATENCY:") +
+                                           strlen("18446744073709551615 ")+1);
                         msgsize = RD_MAX(minlen, msgsize);
                         sendflags |= RD_KAFKA_MSG_F_COPY;
 		} else if (do_seq) {
